Fix field-level asserts
Closes jam1garner/binread#30.
