+++ src/meta_attrs/mod.rs
@@ -7,11 +7,13 @@ pub(crate) use field_level_attrs::FieldLevelAttrs;
 pub(crate) use spanned_value::SpannedValue;
 
 use proc_macro2::TokenStream;
-use crate::compiler_error::SpanError;
-use syn::{Ident, Lit, NestedMeta, Path, Type, Field, Meta, Expr, spanned::Spanned};
+use crate::compiler_error::{CompileError, SpanError};
+use syn::{Expr, Field, Ident, Lit, Meta, NestedMeta, parse::Parse, Path, Type, spanned::Spanned};
 use quote::ToTokens;
 use std::str::FromStr;
 use syn::export::TokenStream2;
+
+use self::parser::MetaList;
 
 #[derive(Debug, Clone)]
 pub struct Assert(pub TokenStream, pub Option<TokenStream>);
@@ -99,4 +101,26 @@ pub enum MagicType {
     Float,
     Bool,
     Verbatim
+}
+
+fn convert_assert<K>(assert: &MetaList<K, Expr>) -> Result<Assert, CompileError>
+    where K: Parse + Spanned,
+{
+    let (cond, err) = match assert.fields[..] {
+        [ref cond] => {
+            (cond, None)
+        }
+        [ref cond, ref err] => {
+            (cond, Some(err))
+        }
+        _ => return SpanError::err(
+            assert.ident.span(),
+            ""
+        ).map_err(Into::into),
+    };
+
+    Ok(Assert(
+        cond.into_token_stream(),
+        err.map(ToTokens::into_token_stream)
+    ))
 }
